// packages/components-library/src/types/theme.ts
export declare type ThemeSettings = {
  fontUrl: string | string[]
}
export declare type ThemeContent = {
  light: ThemeStyles
  dark: ThemeStyles
}
// ...
export declare type ThemeStyles = {
  color: ThemeColor
  border: ThemeBorder
  boxShadow: Record<'xs' | 's' | 'm' | 'brand' | 'l', string>
  font: ThemeFont
  icon: ThemeIcon
  text: ThemeText
  surface: ThemeSurface
  overlay: ThemeOverlay
  typography?: ThemeTypography
  components?: Record<string, any>
}
export interface Theme {
  version: number
  settings: ThemeSettings
  content: Partial<ThemeContent>
}